+++ backend/src/membership/membership.module.ts
@@ -1,12 +1,13 @@
 import { Module, forwardRef } from '@nestjs/common';
 import { PrismaModule } from '../prisma/prisma.module';
 import { PaymentModule } from '../payment/payment.module';
+import { BusinessPolicyModule } from '../business-policy/business-policy.module';
 import { MembershipController } from './membership.controller';
 import { MembershipService } from './membership.service';
 import { MembershipSchedulerService } from './membership-scheduler.service';
 
 @Module({
-  imports: [PrismaModule, forwardRef(() => PaymentModule)],
+  imports: [PrismaModule, forwardRef(() => PaymentModule), BusinessPolicyModule],
   controllers: [MembershipController],
   providers: [MembershipService, MembershipSchedulerService],
   exports: [MembershipService],